api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-chess==1.999
websockets==12.0
pydantic==2.5.2
stockfish==3.28.0
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.25.2